Move http healthcheck to test/http
This is now consistent with the new http/2 and gRPC test servers.
