+++ test/http/serve-healthcheck.go
@@ -1,4 +1,4 @@
-package main
+package http
 
 import (
 	"fmt"
